user add game to card, checkout card, remvoe item from cart

+++ CVGS/CVGS/Controllers/GameController.cs
@@ -17,6 +17,10 @@ namespace CVGS.Controllers
         // GET: Game
         public ActionResult Index()
         {
+            if (Session == null || Session["userID"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             var games = db.Games.Include(g => g.Category).Include(g => g.Category1);
             return View(games.ToList());
         }
@@ -24,6 +28,10 @@ namespace CVGS.Controllers
         // GET: Game/Details/5
         public ActionResult Details(int? id)
         {
+            if (Session == null || Session["userID"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -39,6 +47,10 @@ namespace CVGS.Controllers
         // GET: Game/Create
         public ActionResult Create()
         {
+            if (Session == null || Session["userID"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             ViewBag.categoryID = new SelectList(db.Categories, "categoryID", "categoryName");
 
             var subCategoryList = new SelectList(db.Categories, "categoryID", "categoryName").ToList();
@@ -55,6 +67,10 @@ namespace CVGS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "gameID,title,description,develpoer,publisher,releasedDate,price,categoryID,subCategoryID,createdDate,updatedDate")] Game game)
         {
+            if (Session == null || Session["userID"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (ModelState.IsValid)
             {
                 game.createdDate = DateTime.Now;
@@ -75,6 +91,10 @@ namespace CVGS.Controllers
         // GET: Game/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (Session == null || Session["userID"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -99,6 +119,10 @@ namespace CVGS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "gameID,title,description,develpoer,publisher,releasedDate,price,categoryID,subCategoryID,createdDate,updatedDate")] Game game)
         {
+            if (Session == null || Session["userID"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (ModelState.IsValid)
             {
                 game.updatedDate = DateTime.Now;
@@ -117,6 +141,10 @@ namespace CVGS.Controllers
         // GET: Game/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (Session == null || Session["userID"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -134,6 +162,10 @@ namespace CVGS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (Session == null || Session["userID"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             Game game = db.Games.Find(id);
             db.Games.Remove(game);
             db.SaveChanges();
@@ -147,6 +179,59 @@ namespace CVGS.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+
+        [HttpPost]
+        public ActionResult AddToCart(int gameID)
+        {
+            if (ModelState.IsValid)
+            {
+                if (Session == null || Session["userID"] == null)
+                {
+                    return RedirectToAction("Index", "Home");
+                }
+                CartItem cartItem = new CartItem();
+                cartItem.gameID = gameID;
+
+                int userID = Convert.ToInt32(Session["userID"]);
+                cartItem.userID = userID;
+                cartItem.createdDate = DateTime.Now;
+                cartItem.updatedDate = DateTime.Now;
+
+
+                //end
+                db.CartItems.Add(cartItem);
+                db.SaveChanges();
+                TempData["msg"] = "added to cart";
+                return RedirectToAction("Index");
+            }
+
+            var games = db.Games.Include(g => g.Category).Include(g => g.Category1);
+            return View(games.ToList());
+        }
+
+        public ActionResult AddReview(int? id)
+        {
+            if (Session == null || Session["userID"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            int userID = Convert.ToInt32(Session["userID"]);
+            var res = db.OrderItems.Where(oi => oi.gameID == id && oi.Order.userID == userID);
+            if (res.Count() > 0)
+            {
+                return View();
+            }
+            else
+            {
+                TempData["msg"] = "not owned";
+                return RedirectToAction("Index");
+            }
+        
         }
     }
 }
